+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG
 
-## 0.14.3dev
+## 0.14.5dev
+
+## 0.14.4 (2022-01-07)
+* Fixes an error in the telemetry module
+
+## 0.14.3 (2022-01-06)
+* Improved [anonymous user statistics](https://docs.ploomber.io/en/latest/community/user-stats.html)
 
 ## 0.14.2 (2022-01-03)
 * `PLOOMBER_STATS_ENABLED` environment variable can be used to disable stats